facer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
numpy==1.24.3
torch==1.13.1
torchvision==0.14.1
mediapipe==0.10.7
requests==2.31.0
psutil==5.9.6
Pillow==10.0.1
scikit-image==0.21.0
validators==0.22.0
matplotlib==3.7.2
python-dotenv==1.0.0
scikit-learn==1.3.0